Use host 127.0.0.1 instead of 'google'
The reason therefor is rooted in the sqlserver driver which tries
a dns lookup for the hostname 'google' (which fails). Normally a dns
lookup for the given host totally makes sense to ensure existing hosts.

In combination with Google CloudSQL and the Cloud SQL JDBC connector
for SQL Server a dns lookup is not relevant as the database connection
happens based on the given CloudSQL connection name.

The issue can be bypassed by using 127.0.0.1 which is an IP address,
thus does not trigger a dns lookup.